requirements.txt
pymongo==4.6.1
farm-haystack[mongodb]
